Extract function lowering phase helpers
